+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 ### Fixed
 * Fixed potential concurrency issue on `VideoSourceAdapter`.
 * Fixed a video connection issue on network where DNS 8.8.8.8 is blocked. `ACCESS_NETWORK_STATE` permission is required to discover available DNS on the network.
+* Fixed a case when front camera is missing in the phone.
 
 ## [0.9.1] - 2021-01-08